Separate palette into new file
